+++ Socolin.RabbitMQ.Client/Options/Client/DeliveryMode.cs
@@ -1,0 +1,8 @@
+namespace Socolin.RabbitMQ.Client.Options.Client
+{
+	public enum DeliveryMode
+	{
+		NonPersistent = 1,
+		Persistent = 2
+	}
+}