+++ Socolin.RabbitMQ.Client/Options/Client/RabbitMqServiceOptionsBuilder.cs
@@ -23,6 +23,7 @@ namespace Socolin.RabbitMQ.Client.Options.Client
 		private bool _usePerMessageTtl;
 		private int? _perMessageTTl;
 		private TimeSpan[]? _delaysBetweenRetry;
+		private DeliveryMode? _deliveryMode;
 
 		public RabbitMqServiceOptionsBuilder WithConnectionManager(IRabbitMqConnectionManager connectionManager)
 		{
@@ -105,6 +106,12 @@ namespace Socolin.RabbitMQ.Client.Options.Client
 			return this;
 		}
 
+		public RabbitMqServiceOptionsBuilder WithDeliveryMode(DeliveryMode deliveryMode)
+		{
+			_deliveryMode = deliveryMode;
+			return this;
+		}
+
 		public RabbitMqServiceClientOptions Build()
 		{
 			if (_connectionManager == null)
@@ -127,6 +134,7 @@ namespace Socolin.RabbitMQ.Client.Options.Client
 			if (_usePerMessageTtl)
 				options.PerMessageTtl = new PerMessageTtlOption(_perMessageTTl);
 			options.CustomPipes.AddRange(_customPipeBuilders);
+			options.DeliveryMode = _deliveryMode;
 
 			return options;
 		}