+++ Socolin.RabbitMQ.Client/Pipes/Client/PublishClientPipe.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Threading.Tasks;
+using Socolin.RabbitMQ.Client.Options.Client;
 using Socolin.RabbitMQ.Client.Pipes.Client.Context;
 
 // ReSharper disable ArgumentsStyleLiteral
@@ -8,8 +9,17 @@ namespace Socolin.RabbitMQ.Client.Pipes.Client
 {
 	public class PublishClientPipe : ClientPipe, IMessageClientPipe
 	{
+		private readonly DeliveryMode? _deliveryMode;
+
+		public PublishClientPipe(DeliveryMode? deliveryMode)
+		{
+			_deliveryMode = deliveryMode;
+		}
+
 		public Task ProcessAsync(ClientPipeContextMessage clientPipeContextMessage, ReadOnlyMemory<IClientPipe> pipeline)
 		{
+			if (_deliveryMode.HasValue)
+				clientPipeContextMessage.BasicProperties!.DeliveryMode = (byte) _deliveryMode.Value;
 			clientPipeContextMessage.Channel!.BasicPublish(
 				clientPipeContextMessage.ExchangeName,
 				clientPipeContextMessage.RoutingKey,